fill_ and masked_fill_ support kbk and grad

+++ mindspore/ops/infer/ops_func_impl/cdist.cc
@@ -32,7 +32,7 @@ static inline bool CdistIsValidType(TypeId t) {
 }
 
 ShapeArray CdistFuncImpl::InferShape(const PrimitivePtr &primitive, const InferInfoPtrList &input_infos) const {
-  size_t batch_rank = 0;
+  int64_t batch_rank = 0;
   auto prim_name = primitive->name();
   if (primitive->HasAttr(kBatchRank)) {
     auto value_ptr = primitive->GetAttr(kBatchRank);
